Big Checkboxes (with preview)
This commit atomizes the big checkboxes and converts them into
molecules. It also makes a couple of changes so they no longer require
the form fields and the big-checkboxes are more self-contained.

This has been tested on Voiceover, but may need to be tested with
Jennifer Horan to make sure none of the changes affect JAWS.

This commit adds a small preview form on the Bureau page for reviewing
purposes which will be removed in the next commit.